% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_hardware.R
\name{detect_hardware}
\alias{detect_hardware}
\title{Title}
\usage{
detect_hardware(OS = NULL, TDP = NULL, tracker = FALSE)
}
\arguments{
\item{OS}{NULL par défaut, chaine de caractères pour forcer le nom de l'OS}

\item{TDP}{NULL par défaut, chaine de caractères pour forcer
la valeur du TDP, si elle n'est pas dispo dans la base de données}

\item{tracker}{FALSE par défaut. Si TRUE, le tracker se lance en même temps
que cette fonction. si FALSE, il faut lancer le tracker séparément
ne pas utiliser pour l'instant}
}
\value{
Retourne une liste avec plusieurs paramètres liés au matériel
du PC sur lequel l'analyse tourne
}
\description{
Title
}
\examples{
#ne pas lancer
}
