% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_CPU.R
\name{import_CPU}
\alias{import_CPU}
\title{import_CPU}
\usage{
import_CPU()
}
\value{
A dataframe containing the TDP of AMD and Intel processors.
}
\description{
Function to import TDP data for AMD and Intel processors
from the package's internal files.
}
\examples{
#do not run
}
