% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_rds.R
\name{increment_and_save}
\alias{increment_and_save}
\title{Title}
\usage{
increment_and_save(object_name)
}
\arguments{
\item{object_name}{une chaîne de caractères qui est le nom du fichier rds
à importer et exporter}
}
\value{
un fichier rds stocke localement, test du compteur
}
\description{
Title
}
\examples{
#non
}
