% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_start.R
\name{tracker_start}
\alias{tracker_start}
\title{tracker_start}
\usage{
tracker_start()
}
\value{
retourne un objet date
}
\description{
demarre le tracker, sauvegarde un objet temporel a la date du systeme.
A lancer au debut du programme
}
\examples{
#start_time<-tracker_start()
}
