% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_hardware.R
\name{detect_hardware}
\alias{detect_hardware}
\title{detect_hardware}
\usage{
detect_hardware(OS = NULL, TDP = NULL, tracker = FALSE)
}
\arguments{
\item{OS}{NULL by default. A character string of length 1
to specify the name of the operating system.}

\item{TDP}{NULL by default. A numeric (integer) value of length 1
to specify the TDP value if it is not available in the database.
If the parameter is NULL and the value is not found, the TDP value
is approximated to 15W per core.}

\item{tracker}{FALSE by default. If TRUE, the tracker is launched
along with this function. If FALSE, the tracker must be launched separately.
Do not use at the moment, keep it as FALSE.}
}
\value{
Returns a list with several parameters related to the hardware
of the PC on which the program is running.
}
\description{
Function to detect the hardware on which the program is running.
This function currently works only on Windows operating systems.
Unexpected results may occur with other operating systems at the moment.
}
\examples{
#do not run
#result_hardware <- detect_hardware()
#print(result_hardware)
}
