% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_hardware.R
\name{detect_hardware}
\alias{detect_hardware}
\title{Title}
\usage{
detect_hardware(OS = NULL, TDP = NULL, tracker = FALSE)
}
\arguments{
\item{OS}{NULL par defaut, chaine de caracteres pour forcer le nom de l'OS}

\item{TDP}{NULL par defaut, valeur numerique (entier) pour forcer
la valeur du TDP, si elle n'est pas dispo dans la base de donnees
si le parametre est NULL et que la valeur n'est pas retrouvee, alors
la valeur du TDP est approximee a 15 fois le nombre de coeurs}

\item{tracker}{FALSE par defaut. Si TRUE, le tracker se lance en meme temps
que cette fonction. si FALSE, il faut lancer le tracker separement
ne pas utiliser pour l'instant}
}
\value{
Retourne une liste avec plusieurs parametres lies au materiel
du PC sur lequel l'analyse tourne
}
\description{
Title
}
\examples{
#ne pas lancer
}
