% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_start.R
\name{tracker_start}
\alias{tracker_start}
\title{tracker_start}
\usage{
tracker_start()
}
\value{
Returns a date-time object.
}
\description{
Starts the tracker, saving a timestamp at the current system date and time.
To be called before program execution or at the beginning.
}
\examples{
start_time <- tracker_start()
}
