% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_stop.R
\name{tracker_stop}
\alias{tracker_stop}
\title{Title}
\usage{
tracker_stop(hardware, start_time, path = NULL)
}
\arguments{
\item{hardware}{la liste creee par la fonction 'detect_hardware'}

\item{start_time}{un objet POSIXct, de preference l'objet cree par la
fonction 'tracker_start'}

\item{path}{chemin vers un fichier .rds, utilise pour stocker et
recuperer le temps d'execution total. Laisser vide pour ne produire
qu'une iteration. Renseigner la chaine de}
}
\value{
des choses dans la console
}
\description{
Title
}
\examples{
#tracker_stop(hardware = detect_hardware(),
#start_time = Sys.time(),
#path = "./chemin/vers/fichier.rds")
}
