% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_stop.R
\name{tracker_stop}
\alias{tracker_stop}
\title{tracker_stop}
\usage{
tracker_stop(hardware, start_time, carbon_intensity = 0.06207, path = NULL)
}
\arguments{
\item{hardware}{The list created previously by the 'detect_hardware' function.}

\item{start_time}{A POSIXct object, preferably the one created by the 'tracker_start' function.}

\item{carbon_intensity}{Electricity Generation Emissions Factor. Default is France July 2023 = 0.06207 (source: https://www.carbonfootprint.com/international_electricity_factors.html)}

\item{path}{Path to a .rds file, used to store and retrieve the total execution time. Leave empty to only produce a single iteration. Provide the file path to an existing or non-existing .rds file to accumulate execution time from previous and subsequent runs.}
}
\value{
Outputs a report to the console, and if the 'path' argument is provided, saves a .rds file at that path.
}
\description{
Calculates the energy consumption and carbon footprint of the program from the moment the tracker was started. It can also accumulate these indicators depending on how many times the program has been run. To be called after program execution or at the end.
}
\examples{
# tracker_stop(hardware = detect_hardware(),
# start_time = Sys.time(),
# path = "./path/to/file.rds")
}
