% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_stop.R
\name{tracker_stop}
\alias{tracker_stop}
\title{Title}
\usage{
tracker_stop(hardware, start_time)
}
\arguments{
\item{hardware}{la liste creee par la fonction 'detect_hardware'}

\item{start_time}{un objet POSIXct, de preference l'objet cree par la
fonction 'tracker_start'}
}
\value{
des choses dans la console
}
\description{
Title
}
\examples{
#tracker_stop<-function(start = tracker)
}
